requirements.txt

numpy==1.24.1

# dev
neptune-client==0.16.15
requests
loguru==0.6.0
pydantic==1.10.2
tqdm==4.64.1
python-dotenv==0.19.2

# dataviz
matplotlib==3.6.2

